Use pcall instead of task.spawn to invoke notifications.
-- Threads --

Using pcall keeps the invocation of a notification on the same thread.
This will likely be important later as we begin thinking more about how
Rx interacts with threads.

Moreover, relative to a normal function call, pcall is likely either
faster than or the same as task.spawn, so it's not worth preferring one
over the other in regards to speed (speed shouldn't even be considered
at this point anyway).

-- Errors --

In general, errors in Rx are passed as values through the OnError
notification. These value-type errors can be considered one class of
error.

There is another class of error that is more akin to a "panic". That is,
a problem with the implementation rather than from the user. pcall is
used to "recover" from these panics that occur.

An error is "observable" when it is sent through an OnError
notification. User errors should definitely be observable. Whether
panics should be observable remains to be decided.

For now, to deal with errors, we attempt to invoke the OnError
notification. If that manages to panic, it is emitted as a warning.
That is, panics from OnNext/OnComplete are observable, while panics
from OnError are not.